Add missing copyright headers
Previous revisions that does not contain such headers will remain as non-licensed, all rights reserved. As long as the usages of such files refer to this commit's revision, no problem would happen.
If someone does need the (deleted) files that does not have the headers, we would need to rebase the whole git history.